# api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Heimdallr API
  version: 1.0.0
  description: API for managing applications, releases, compliance reports, servers, agents, providers, automations, and automation jobs.
servers:
  - url: /api
    description: API base path
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/application/{application_id}/release:
    get:
      tags:
        - Release
      summary: List releases for an application
      operationId: listReleases
      parameters:
        - $ref: "#/components/parameters/ApplicationIDPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Releases returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Release
      summary: Create release for an application
      operationId: createRelease
      parameters:
        - $ref: "#/components/parameters/ApplicationIDPath"
        - name: upsert
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true, update an existing release with the same version instead of returning a conflict
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseCreateRequest"
      responses:
        "201":
          description: Release created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/server:
    get:
      tags:
        - Server
      summary: List servers
      operationId: listServers
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: agent_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
          description: Return only servers where the given agent is installed
      responses:
        "200":
          description: Servers returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Server
      summary: Create server
      operationId: createServer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerCreateRequest"
      responses:
        "201":
          description: Server created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/server/{server_id}:
    get:
      tags:
        - Server
      summary: Get server by ID
      operationId: getServer
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
      responses:
        "200":
          description: Server returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerWithRelationsDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags:
        - Server
      summary: Register agents on an existing server
      operationId: updateServer
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerUpdateRequest"
      responses:
        "200":
          description: Server updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerWithRelationsDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/server/{server_id}/job:
    get:
      tags:
        - Server
      summary: List automation jobs associated with a server
      operationId: listServerJobs
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Associated jobs returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerJobListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Server
      summary: Associate an existing automation job with a server
      operationId: associateServerJob
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerJobAssociateRequest"
      responses:
        "201":
          description: Job associated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/server/{server_id}/job/{job_id}:
    delete:
      tags:
        - Server
      summary: Dissociate an automation job from a server
      operationId: dissociateServerJob
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
        - $ref: "#/components/parameters/JobID"
        - name: automation_id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Job dissociated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/server/{server_id}/release:
    get:
      tags:
        - Server
      summary: List releases associated with a server
      operationId: listServerReleases
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Associated releases returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerReleaseListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Server
      summary: Associate an existing release with a server
      operationId: associateServerRelease
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerReleaseAssociateRequest"
      responses:
        "201":
          description: Release associated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/server/{server_id}/release/{release_id}:
    delete:
      tags:
        - Server
      summary: Dissociate a release from a server
      operationId: dissociateServerRelease
      parameters:
        - $ref: "#/components/parameters/ServerIDPath"
        - name: release_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Release dissociated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API token passed as a Bearer token in the Authorization header
    SessionCookie:
      type: apiKey
      in: cookie
      name: heimdallr_session
      description: HttpOnly browser session cookie
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    JobID:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
    ServerIDPath:
      name: server_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ApplicationIDPath:
      name: application_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UUID:
      type: string
      format: uuid
      x-go-type: uuid.UUID
      x-go-type-import:
        path: github.com/google/uuid
        name: uuid
    URL:
      type: string
      format: uri
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: invalid request body
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    JobStatus:
      type: string
      enum:
        - started
        - skipped
        - success
        - failed
...
    ServerMetadata:
      type: object
      properties: {}
    ServerRelationSummary:
      type: object
      required:
        - agent_count
        - job_count
        - release_count
      properties:
        agent_count:
          type: integer
          minimum: 0
        job_count:
          type: integer
          minimum: 0
        release_count:
          type: integer
          minimum: 0
    Server:
      type: object
      required:
        - id
        - hostname
        - metadata
        - operating_system
        - hypervisor
        - location
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        hostname:
          type: string
          minLength: 1
          maxLength: 255
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
        operating_system:
          type: string
        hypervisor:
          type: string
        location:
          type: string
    ServerListItem:
      allOf:
        - $ref: "#/components/schemas/Server"
        - type: object
          required:
            - relations
          properties:
            relations:
              $ref: "#/components/schemas/ServerRelationSummary"
    ServerWithRelations:
      allOf:
        - $ref: "#/components/schemas/Server"
        - type: object
          required:
            - relations
          properties:
            relations:
              $ref: "#/components/schemas/ServerRelationSummary"
    ServerCreateRequest:
      type: object
      required:
        - hostname
      properties:
        hostname:
          type: string
          minLength: 1
          maxLength: 255
        operating_system:
          type: string
          maxLength: 255
        hypervisor:
          type: string
          maxLength: 255
        location:
          type: string
          maxLength: 255
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
        agent_ids:
          type: array
          items:
            $ref: "#/components/schemas/UUID"
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentCreateRequest"
    ServerUpdateRequest:
      type: object
      properties:
        agent_ids:
          type: array
          items:
            $ref: "#/components/schemas/UUID"
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentCreateRequest"
    ServerDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Server"
    ServerWithRelationsDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/ServerWithRelations"
    ServerListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServerListItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    AgentCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        type:
          type: string
          maxLength: 255
        version:
          type: string
          maxLength: 255
        metadata:
          $ref: "#/components/schemas/ServerMetadata"
...
    ServerJobAssociation:
      type: object
      required:
        - job_id
        - automation_id
        - automation
        - provider
        - status
        - location
        - url
      properties:
        job_id:
          type: string
          minLength: 1
        automation_id:
          $ref: "#/components/schemas/UUID"
        automation:
          type: string
        provider:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        location:
          type: string
        url:
          $ref: "#/components/schemas/URL"
    ServerJobAssociateRequest:
      type: object
      required:
        - job_id
        - automation_id
      properties:
        job_id:
          type: string
          minLength: 1
        automation_id:
          $ref: "#/components/schemas/UUID"
    ServerJobListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServerJobAssociation"
        pagination:
          $ref: "#/components/schemas/Pagination"
    ServerReleaseAssociation:
      type: object
      required:
        - release_id
        - application_id
        - application
        - version
        - commit_sha
        - branch
      properties:
        release_id:
          $ref: "#/components/schemas/UUID"
        application_id:
          $ref: "#/components/schemas/UUID"
        application:
          type: string
        version:
          type: string
        commit_sha:
          type: string
        branch:
          type: string
    ServerReleaseAssociateRequest:
      type: object
      required:
        - release_id
        - application_id
      properties:
        release_id:
          $ref: "#/components/schemas/UUID"
        application_id:
          $ref: "#/components/schemas/UUID"
    ServerReleaseListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServerReleaseAssociation"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    Release:
      type: object
      required:
        - id
        - application_id
        - application
        - version
        - commit_sha
        - pipeline_url
        - branch
        - created_at
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        application_id:
          $ref: "#/components/schemas/UUID"
        application:
          type: string
        version:
          type: string
        commit_sha:
          type: string
        pipeline_url:
          $ref: "#/components/schemas/URL"
        branch:
          type: string
        created_at:
          type: string
          format: date-time
    ReportSummary:
      type: object
      required:
        - type
        - status
        - count
      properties:
        type:
          type: string
        status:
          type: string
        count:
          type: integer
          minimum: 0
    ComplianceSummary:
      type: object
      required:
        - total_reports
        - successful_reports
        - failed_reports
        - started_reports
        - success_rate
        - by_type
      properties:
        total_reports:
          type: integer
          minimum: 0
        successful_reports:
          type: integer
          minimum: 0
        failed_reports:
          type: integer
          minimum: 0
        started_reports:
          type: integer
          minimum: 0
        success_rate:
          type: number
          format: double
        by_type:
          type: array
          items:
            $ref: "#/components/schemas/ReportSummary"
    ReleaseListItem:
      allOf:
        - $ref: "#/components/schemas/Release"
        - type: object
          required:
            - compliance
          properties:
            compliance:
              $ref: "#/components/schemas/ComplianceSummary"
...
    ReleaseCreateRequest:
      type: object
      required:
        - version
      properties:
        version:
          type: string
          minLength: 1
          maxLength: 255
        commit_sha:
          type: string
          maxLength: 255
        pipeline_url:
          $ref: "#/components/schemas/URL"
        branch:
          type: string
          maxLength: 255
    ReleaseDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Release"
    ReleaseListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReleaseListItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
...